Mapbox funcionando, se muestra el mapa correctamente. Se creó una carpeta plugins donde se encuentra init_mapbox.js
